cast mProbeInterval to long

+++ dcc/src/ChannelProber.cpp
@@ -38,7 +38,7 @@ ChannelProber::ChannelProber(string ifname, double probeInterval, boost::asio::i
 	mProbeInterval = probeInterval;
 	mIfname = ifname;
 	mIoService = io;
-	mTimer = new boost::asio::deadline_timer(*mIoService, boost::posix_time::millisec(probeInterval * 1000));
+	mTimer = new boost::asio::deadline_timer(*mIoService, boost::posix_time::millisec(long(probeInterval * 1000)));
 }
 
 ChannelProber::~ChannelProber() {
@@ -193,7 +193,7 @@ void ChannelProber::probe(const boost::system::error_code &ec) {
 			NLM_F_DUMP);
 	nl_recvmsgs_default(mSocket);
 
-	mTimer->expires_from_now(boost::posix_time::millisec(mProbeInterval * 1000));
+	mTimer->expires_from_now(boost::posix_time::millisec(long(mProbeInterval * 1000)));
 	mTimer->async_wait(boost::bind(&ChannelProber::probe, this, boost::asio::placeholders::error));
 }
